Added initial support for recording named tracks.
New tracks can be created and stored in the database.  All points logged are labeled with the current track.  A basic name, and description information is associated with the track.  The implementation is still incomplete.

Dropped @Override directives on interface implements.

Moved more of the static text in layouts to the strings resource.

Fixed possible threading issue with using SQLite in Logger upload thread.