Added pretty print to GPSList activity.
When quering locations from the Content Provider, a new query parameter allows the format of the data to be presented differently in the cursor.

Added a projection map for the tracks table in the Content Provider.

TrackListActivity now sends in a Uri to the track instead of track row id.

Fixed a bug in the Content Provider with update/delete not including a space in the AND WHERE clause.